Rename entryModelData to proteomesEntryModelData as mocks must be globally unique.
